Removed dependency lookup from Laravel collector, and created new Composer collector that it invokes instead.
